feat: add new landing page and move original homepage to /app
- Created a new professional landing page at `/` with logo, intro text, and vision for Ovotime
- Moved the original homepage to `/app` and updated component import paths
- Added a prominent link from the landing page to the tools/dashboard
